% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_utils.R
\name{close_log_file}
\alias{close_log_file}
\title{Close and clear the shared log file connection if present.}
\usage{
close_log_file()
}
\description{
Close and clear the shared log file connection if present.
}
