% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_utils.R
\name{log_debug}
\alias{log_debug}
\title{Write debug-level messages with a consistent prefix.}
\usage{
log_debug(msg)
}
\arguments{
\item{msg}{Character string to log.}
}
\description{
Write debug-level messages with a consistent prefix.
}
