% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PH_PostProcessing_andAnalysis.R
\name{apply_all_clustering_methods}
\alias{apply_all_clustering_methods}
\title{Apply all clustering approaches to a Seurat object}
\usage{
apply_all_clustering_methods(
  seurat_obj,
  dataset_name,
  assay,
  bdm_matrix = NULL,
  sdm_matrix = NULL,
  landscape_matrix = NULL,
  run_kmeans_clustering = TRUE,
  run_hierarchical_ph_clustering = TRUE,
  run_spectral_clustering = TRUE,
  SRA_col = "orig.ident",
  Tissue_col = "Tissue",
  Approach_col = "Approach"
)
}
\description{
Apply all clustering approaches to a Seurat object
}
