% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_iteration_functions.R
\name{bootstrap_aggregated_landscape_curve}
\alias{bootstrap_aggregated_landscape_curve}
\title{Bootstrap an aggregated landscape curve}
\usage{
bootstrap_aggregated_landscape_curve(
  individual_curves,
  grid_points,
  n_bootstrap = 100
)
}
\arguments{
\item{individual_curves}{List of numeric vectors representing individual curves.}

\item{grid_points}{Number of grid points.}

\item{n_bootstrap}{Number of bootstrap samples.}
}
\value{
List with mean, lower, upper curves and all bootstrap samples.
}
\description{
Bootstrap an aggregated landscape curve
}
\examples{
bootstrap_aggregated_landscape_curve(list(c(0,1,0)), 3, n_bootstrap = 2)
}
