% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_iteration_functions.R
\name{bootstrap_null_stats_landscape}
\alias{bootstrap_null_stats_landscape}
\title{Bootstrap null statistics for landscape summaries}
\usage{
bootstrap_null_stats_landscape(
  landscape_list,
  n_bootstrap = 50,
  grid_points = 500
)
}
\arguments{
\item{landscape_list}{List of landscape objects.}

\item{n_bootstrap}{Number of bootstrap samples.}

\item{grid_points}{Number of grid points for evaluation.}
}
\value{
List containing null effect and KS statistics.
}
\description{
Bootstrap null statistics for landscape summaries
}
\examples{
bootstrap_null_stats_landscape(list(list(dim0 = c(0,1), dim1 = c(0,0))),
                               n_bootstrap = 2, grid_points = 5)
}
