% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_iteration_functions.R
\name{compute_aggregated_landscape_curve}
\alias{compute_aggregated_landscape_curve}
\title{Aggregate multiple landscape curves}
\usage{
compute_aggregated_landscape_curve(
  landscape_list_group,
  grid = seq(0, 1, length.out = grid_points)
)
}
\arguments{
\item{landscape_list_group}{List of landscape objects.}

\item{grid}{Numeric vector of grid points.}
}
\value{
Mean aggregated landscape curve.
}
\description{
Aggregate multiple landscape curves
}
\examples{
compute_aggregated_landscape_curve(list(list(dim0 = c(0,1), dim1 = c(0,0))),
                                  seq(0, 1, length.out = 3))
}
