% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_iteration_functions.R
\name{compute_null_stats_aggregated_landscapes}
\alias{compute_null_stats_aggregated_landscapes}
\title{Bootstrap null stats for aggregated landscapes}
\usage{
compute_null_stats_aggregated_landscapes(
  pd_list,
  landscape_list,
  tau_vals,
  n_bootstrap = 50,
  num_cores = 8,
  grid_points = 500
)
}
\arguments{
\item{pd_list}{List of persistence diagrams.}

\item{landscape_list}{List of landscape objects.}

\item{tau_vals}{Numeric vector of tau values.}

\item{n_bootstrap}{Number of bootstrap samples.}

\item{num_cores}{Number of CPU cores.}

\item{grid_points}{Number of grid points.}
}
\value{
List with null effect and KS statistics.
}
\description{
Bootstrap null stats for aggregated landscapes
}
\examples{
compute_null_stats_aggregated_landscapes(list(matrix(c(0,0,1), ncol=3)),
                                        list(list(dim0=c(0,1), dim1=c(0,0))),
                                        c(0,1))
}
