% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_iteration_functions.R
\name{compute_null_stats_individual_landscapes}
\alias{compute_null_stats_individual_landscapes}
\title{Bootstrap null stats for individual landscapes}
\usage{
compute_null_stats_individual_landscapes(
  landscape_list,
  n_bootstrap = 50,
  grid_points = 500
)
}
\arguments{
\item{landscape_list}{List of landscape objects.}

\item{n_bootstrap}{Number of bootstrap iterations.}

\item{grid_points}{Number of grid points.}
}
\value{
List with null effect and KS statistics or `NULL` if insufficient data.
}
\description{
Bootstrap null stats for individual landscapes
}
\examples{
compute_null_stats_individual_landscapes(list(list(dim0 = c(0,1), dim1 = c(0,0))),
                                        n_bootstrap = 2, grid_points = 5)
}
