% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PH_PostProcessing_andAnalysis.R
\name{generate_visualizations_for_iteration}
\alias{generate_visualizations_for_iteration}
\title{Generate visualizations and heatmaps for a single iteration}
\usage{
generate_visualizations_for_iteration(
  seurat_obj,
  dataset_name,
  assay,
  bdm_matrix = NULL,
  sdm_matrix = NULL,
  landscape_matrix = NULL,
  metadata = NULL,
  plots_folder = "plots",
  run_visualizations = TRUE,
  run_sample_level_heatmap = TRUE,
  SRA_col = "orig.ident",
  Tissue_col = "Tissue",
  Approach_col = "Approach"
)
}
\description{
Generate visualizations and heatmaps for a single iteration
}
