% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_iteration_functions.R
\name{integrated_diff}
\alias{integrated_diff}
\title{Integrated squared difference between curves}
\usage{
integrated_diff(curve_diff, tau_vals)
}
\arguments{
\item{curve_diff}{Numeric vector of curve differences.}

\item{tau_vals}{Numeric vector of tau values.}
}
\value{
Scalar integrated distance.
}
\description{
Integrated squared difference between curves
}
\examples{
integrated_diff(c(0, 1, 0), c(0, 0.5, 1))
}
