% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PH_PostProcessing_andAnalysis.R
\name{perform_spectral_clustering}
\alias{perform_spectral_clustering}
\title{Perform spectral clustering on a distance matrix}
\usage{
perform_spectral_clustering(distance_matrix, k)
}
\description{
Perform spectral clustering on a distance matrix
}
