% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_comparison.R
\name{run_cluster_comparison}
\alias{run_cluster_comparison}
\title{Compare clustering results across methods.}
\usage{
run_cluster_comparison(
  data_iterations,
  results_folder,
  dataset_tag = "dataset",
  run_comparative_metrics = TRUE,
  include_silhouette = FALSE,
  SRA_col = "SRA",
  verbose = TRUE
)
}
\description{
This function assumes clustering has already been performed on the
provided Seurat objects.  It collects the clustering columns and
produces comparative metrics and plots.
}
